Refactor SchemaLoader into EntityManager, remove SchemaLoader dependency
